requirements.txt
asgiref==3.3.1
autopep8==1.5.6
coverage==5.5
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==8.0.0
model-bakery==1.2.1
pycodestyle==2.7.0
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Unidecode==1.2.0
